Refactor [ Filter / FetchMovieList 분리 ] - 요구사항에따라 Filterring하는 조건이 얼마든지 달라질 수 있다고 봄. - 때문에 이 부분은 만약 기획자랑 상의한다면 유동적인 Usecase가 될수 있고 독립적으로 관리 가능해야 한다고 생각했음.
