+++ 01_presentaion/src/main/java/com/ppizil/presentaion/model/movielist/MovieCategoryModel.kt
@@ -1,6 +1,6 @@
-package com.ppizil.presentaion.movielist
+package com.ppizil.presentaion.model.movielist
 
-import com.ppizil.presentaion.movielist.MovieModel
+import com.ppizil.presentaion.movielist.MovieCategories
 
 data class MovieCategoryModel(
     val category: MovieCategories,
