+++ 01_presentaion/src/main/java/com/ppizil/presentaion/model/movielist/ParamMovieModel.kt
@@ -1,0 +1,7 @@
+package com.ppizil.presentaion.model.movielist
+
+data class MovieModel(
+    val language : String?,
+    val region : String?
+)
+
