+++ 01_presentaion/src/main/java/com/ppizil/presentaion/model/movielist/ViewMovieModel.kt
@@ -1,4 +1,6 @@
-package com.ppizil.presentaion.movielist
+package com.ppizil.presentaion.model.movielist
+
+import com.ppizil.presentaion.movielist.MovieCategories
 
 data class ViewMovieModel(
     val id: Long?,
