+++ 01_presentaion/src/main/java/com/ppizil/presentaion/movielist/MovieListState.kt
@@ -1,5 +1,7 @@
 package com.ppizil.presentaion.movielist
 
+import com.ppizil.presentaion.model.movielist.MovieCategoryModel
+
 sealed class MovieListState {
     data class Success(val movies: List<MovieCategoryModel?>) : MovieListState()
     object Dismiss : MovieListState()
