--- 01_presentaion/src/main/java/com/ppizil/presentaion/movielist/MovieMapper.kt
@@ -1,19 +1,0 @@
-package com.ppizil.presentaion.movielist
-
-import com.ppizil.domain.model.ParamMovieEntity
-import com.ppizil.domain.model.ResultMovieEntity
-
-
-fun ResultMovieEntity.mapToPresent() = ViewMovieModel(
-    id = id,
-    title = title,
-    contents = desc,
-    imageUrl = imageUrl,
-    category = category.mapToCategory(),
-)
-
-
-fun MovieModel.mapToDomain() = ParamMovieEntity(
-    language = language,
-    region = region
-)
