--- 01_presentaion/src/main/java/com/ppizil/presentaion/movielist/ParamMovieModel.kt
@@ -1,10 +1,0 @@
-package com.ppizil.presentaion.movielist
-
-import com.ppizil.domain.model.ParamMovieEntity
-import com.ppizil.domain.model.ResultMovieEntity
-
-data class MovieModel(
-    val language : String?,
-    val region : String?
-)
-
