+++ 02_domain/src/main/java/com/ppizil/domain/di/DomainModule.kt
@@ -25,4 +25,8 @@ object DomainModule {
         createMoveListParam,
         repo
     )
+
+    @Provides
+    @Singleton
+    fun provideFilterringMovieList() : FilterMovieContents = FilterMovieContentsImpl()
 }
