+++ 02_domain/src/main/java/com/ppizil/domain/model/ResultFilterredEntity.kt
@@ -1,0 +1,6 @@
+package com.ppizil.domain.model
+
+data class ResultFilterredEntity(
+    val category: Int,
+    val movies: List<ResultMovieEntity>
+)
