+++ 02_domain/src/main/java/com/ppizil/domain/usecase/FilterMovieContents.kt
@@ -1,0 +1,8 @@
+package com.ppizil.domain.usecase
+
+import com.ppizil.domain.model.ResultFilterredEntity
+import com.ppizil.domain.model.ResultMovieEntity
+
+interface FilterMovieContents {
+    suspend operator fun invoke(items : List<ResultMovieEntity>) : List<ResultFilterredEntity>
+}
