+++ app/src/main/java/com/ppizil/cooflix/movielist/MovieContentsAdapter.kt
@@ -8,7 +8,7 @@ import com.bumptech.glide.RequestManager
 import com.ppizil.cooflix.base.BaseListAdapter
 import com.ppizil.cooflix.databinding.ItemMovieContentsBinding
 import com.ppizil.cooflix.movielist.holder.MovieContentsHolder
-import com.ppizil.presentaion.movielist.ViewMovieModel
+import com.ppizil.presentaion.model.movielist.ViewMovieModel
 
 class MovieContentsAdapter(
     private val requestManager: RequestManager
