+++ app/src/main/res/layout/item_movie_category_section.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="item"
-            type="com.ppizil.presentaion.movielist.MovieCategoryModel" />
+            type="com.ppizil.presentaion.model.movielist.MovieCategoryModel" />
 
         <variable
             name="requestManager"
