+++ app/src/main/res/layout/item_movie_contents.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="item"
-            type="com.ppizil.presentaion.movielist.ViewMovieModel" />
+            type="com.ppizil.presentaion.model.movielist.ViewMovieModel" />
 
         <variable
             name="requestManager"
